Cleanup filesystem creds more optimistically
We can delete the creds on the filesystem (currently just docker) as soon as goreleaser finishes. We still call cleanup on exit to ensure we always cleanup.

+++ scripts/release.sh
@@ -2,13 +2,22 @@
 
 set -e
 
+CLEAN=0
+
 function cleanup {
+  if [ "$CLEAN" -eq 1 ]; then
+    return
+  fi
+
   # delete docker creds
   set +e
   docker logout $DOCKER_REGISTRY
   docker logout $GCR_REGISTRY
   set -e
   rm -rf "$DOCKER_CONFIG"
+
+  # we have cleaned
+  CLEAN=1
 }
 trap cleanup EXIT
 
@@ -19,5 +28,6 @@ echo $DOCKER_HUB_TOKEN | docker login -u $DOCKER_HUB_USER --password-stdin $DOCK
 echo $GOOGLE_CREDS | docker login -u $GCR_USER --password-stdin $GCR_REGISTRY
 
 GOOGLE_APPLICATION_CREDENTIALS=<(echo "$GOOGLE_CREDS") goreleaser release --rm-dist
+cleanup
 scripts/publish-deb.sh
 scripts/publish-rpm.sh
